+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 - #task_number - short task description
+- #136 - Checking functionality ReviewImage Rest
+- #146 - created a log starter
+- #185 - Add Spotbugs
 - #186 - Add CheckStyle
 
 ## [1.4.0] - 2.04.2024